add precommit
