+++ .golangci.yml
@@ -37,7 +37,6 @@ linters:
   - noctx
   - prealloc
   - predeclared
-  - revive
   - rowserrcheck
   - sqlclosecheck
   - staticcheck
@@ -52,6 +51,7 @@ linters:
   - wastedassign
   - whitespace
   disable:
+  - revive # FIXME re-enable when found a way to ignore exported
   - cyclop
   - exhaustivestruct
   - funlen
@@ -95,18 +95,11 @@ linters-settings:
   gomnd:
     settings:
       mnd:
-        # don't include the "operation" and "assign"
         checks: argument,case,condition,return
   lll:
     line-length: 140
   maligned:
     suggest-new: true
-  nolintlint:
-    allow-leading-space: true # don't require machine-readable nolint directives (i.e. with no leading space)
-    allow-unused: false # report any unused nolint directives
-    require-explanation: false # don't require an explanation for nolint directives
-    require-specific: false # don't require nolint directives to be specific about which linter is being skipped
-
 
 issues:
   exclude-rules:
