+++ README.md
@@ -7,6 +7,7 @@
         <a href="https://codecov.io/gh/Iridaceae/iridaceae"><img alt="Codecov" src="https://img.shields.io/codecov/c/gh/Iridaceae/iridaceae?style=flat-square&token=qGdkowt7ki"></a>
         <a href="https://github.com/Iridaceae/iridaceae/issues"><img alt="GitHub issues" src="https://img.shields.io/github/issues/Iridaceae/iridaceae?style=flat-square&color=important"></a>
         <img alt="GitHub commit checks state" src="https://img.shields.io/github/checks-status/Iridaceae/iridaceae/a29703a1367977d2867167fda8c4146aea6cd58e?style=flat-square">
+        <a href="https://github.com/pre-commit/pre-commit"><img src="https://img.shields.io/badge/pre--commit-enabled-brightgreen?style=flat-square&logo=pre-commit&logoColor=white" alt="pre-commit" style="max-width:100%;"></a>
     </p>
     <br>
     <strong>A general purpose discord bot that focuses on readability and developer-friendly<br></strong>
