+++ go.mod
@@ -15,4 +15,5 @@ require (
 	github.com/stretchr/testify v1.7.0
 	github.com/zekroTJA/timedmap v1.3.1
 	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c
 )
