+++ internal/database/database.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"strconv"
 
-	"github.com/Iridaceae/iridaceae/pkg/helpers"
+	helpers2 "github.com/Iridaceae/iridaceae/internal/helpers"
 
 	"github.com/Iridaceae/iridaceae/pkg/log"
 
@@ -14,7 +14,7 @@ import (
 )
 
 func init() {
-	_ = helpers.LoadGlobalEnv()
+	_ = helpers2.LoadGlobalEnv()
 	mUser := os.Getenv("IRIS_MONGO_USER")
 	mPass := os.Getenv("IRIS_MONGO_PASS")
 	mDBName := os.Getenv("IRIS_MONGO_DBNAME")
