+++ internal/testutils/cbor/examples/makefile
@@ -1,10 +1,10 @@
 all: genLogJSON genLogCBOR
 
 genLogJSON: genLog.go
-	go build -o genLogJSON genLog.go 
+	go build -o genLogJSON genLog.go
 
 genLogCBOR: genLog.go
-	go build -tags binary_log -o genLogCBOR genLog.go 
+	go build -tags binary_log -o genLogCBOR genLog.go
 
 clean:
 	rm -f genLogJSON genLogCBOR
