--- pkg/configmanager/yamlparser_test.go
@@ -1,14 +1,0 @@
-package configmanager
-
-import (
-	"bytes"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-var testYamlParser = &YamlParser{}
-
-func TestYamlParser_Marshal(t *testing.T) {
-	assert.Nil(t, testYamlParser.Marshal(&bytes.Buffer{}, GetIrisConfig()))
-}
